Warn for `inconsistent_struct_constructor`
See: https://rust-lang.github.io/rust-clippy/master/index.html#inconsistent_struct_constructor

+++ crates/typst-library/src/text/smartquote.rs
@@ -410,7 +410,7 @@ cast! {
 
         value.finish(&keys)?;
 
-        Self { single, double }
+        Self { double, single }
     },
     value: SmartQuoteSet => Self {
         double: Smart::Custom(value),
